llm-sandbox/01-first-api-call/pyproject.toml
[project]
name = "01-first-api-call"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.47.0",
    "python-dotenv>=1.2.1",
]
